Merge from Havner/fork (#567)
* Split target controller with emulated controller

Target is a physical one, emulated is emulated one. Don't mix them. This fixes
an issue where the code tried to unplug emulated controller on shutdown when no
physical was connected causing a crash. It also prevents any operations on a
physical one to happen on emulated one.

Fix all the code so the targetController can be null.

* Truly make the XBox360 target controller the default one

The issue was that HIDmode was not set before the service start and it defaulted
to 0, which was DS4, using its assets for a case when there was no physical
controller connected.

* Simplify the neptuneController mute check

* fix an issue with sensor combobox selection on disabled items

+1 more

---------

Co-authored-by: Lukasz Pawelczyk <havner@gmail.com>