(PoC) Prevent full screen games from minimizing when using QuickTools menu.
Reference: https://stackoverflow.com/questions/66512492/c-sharp-wpf-overlay-for-fullscreen-applications

Implementation of WindowSinker using HWND_TOPMOST flag instead of HWND_BOTTOM for QuickTools menu. Testing indicates that most games are no longer minimizing when bringing the QuickTools menu up. However, because there is no focus on the QuickTools menu, you can't change options. Interacting with the QuickTools menu will minimize the game. Possible solution is to capture controller actions independently and send them to the QuickTools window ?

Note: some games like Elden Ring will continue to minimize even when using this solution. I believe it's because the game is using exclusive fullscreen. My research leads me to think we won't be able to fix this with this approach. We will possibly need a different technology such as Metro apps or a driver-level overlay.

This is a proof-of-concept for research and testing purposes, therefore it should not be considered as a working solution.