Update ADLX_Wrapper.dll (#239)
This is the same code as the previous PR (#236) but I noticed the filesize was bigger.
Inspecting it revealed it was a DLL (GUI) (whatever that is) versus a DLL (console).
I was also getting access violation exceptions that I would not expect.

I re-compiled it from the VS Developer Command Prompt instead of the GUI.
The filesize now more closely matches the previous file and it no longer reports as DLL (GUI).
I have also not been able to reproduce the access violation exceptions..