Feature: Add controls for Intel Endurance Gaming
Give the user the option to control and configure Intel's Endurance Gaming in both Profile and Quick Menu.

By default "Super Batter Saver" turns on Endurance Gaming with "Efficiency, which reduces the frame rate to 30 fps.

This gives the user granular control the power saving of Inte's tuning based on FPS, yet still target various fps rates such as 45 in balanced and 60 in performance.