Feature: Add Intel Endurance Gaming to Power profile options
Give the user the option to control and configure Intel's Endurance Gaming in both Profile and Quick Menu.

By default "Super Batter Saver" turns on Endurance Gaming with "Efficiency, which reduces the frame rate to 30 fps.

This gives the user granular control the power saving of Inte's tuning based on FPS.

Signed-off-by: Bassem Mohsen <bassem.nomany@gmail.com>